Move server config from internal/config to pkg/server/config
This improves the package structure by placing server configuration
alongside the server implementation, making the API more discoverable
and the package structure more intuitive.
