+++ sonic-gnmi-standalone/cmd/server/main.go
@@ -7,8 +7,8 @@ import (
 	"syscall"
 
 	"github.com/golang/glog"
-	"github.com/sonic-net/sonic-gnmi/sonic-gnmi-standalone/internal/config"
 	"github.com/sonic-net/sonic-gnmi/sonic-gnmi-standalone/pkg/server"
+	"github.com/sonic-net/sonic-gnmi/sonic-gnmi-standalone/pkg/server/config"
 )
 
 func main() {
